+++ extlibs/vili/src/DataNode.cpp
@@ -79,7 +79,7 @@ namespace vili
             throw aube::ErrorHandler::Raise("Vili.Vili.DataNode.WrongBoolSet", {{"path", getNodePath()},{"type", Types::dataTypeToString(m_dataType)}});
     }
 
-    void DataNode::set(unsigned long long var)
+    void DataNode::set(uint64_t var)
     {
         if (m_dataType == DataType::Long)
             m_data = var;
@@ -112,7 +112,7 @@ namespace vili
         this->set(var);
     }
 
-    void DataNode::operator=(unsigned long long var)
+    void DataNode::operator=(uint64_t var)
     {
         this->set(var);
     }
@@ -131,7 +131,7 @@ namespace vili
             else if (m_dataType == DataType::Bool)
                 m_data = (rawData == "True" ? true : false);
             else if (m_dataType == DataType::Long)
-                m_data = stoull(rawData);
+                m_data = static_cast<uint64_t>(stoull(rawData));
         }
         else
             throw aube::ErrorHandler::Raise("Vili.Vili.DataNode.WrongAutoset", {{"data", rawData},{"path", getNodePath()}});
@@ -148,7 +148,7 @@ namespace vili
         if (m_dataType == DataType::Bool)
             return (std::get<bool>(m_data) ? "True" : "False");
         if (m_dataType == DataType::Long)
-            return std::to_string(std::get<unsigned long long>(m_data));
+            return std::to_string(std::get<uint64_t>(m_data));
         return "?TypeError?";
     }
